+++ src/app/api/v1/oidc/[...oidc]/route.ts
@@ -10,7 +10,11 @@ import { getProvider } from "@/lib/oidc/provider";
 import { IncomingMessage, ServerResponse } from "http";
 import { Socket } from "net";
 import { normalizeProviderPath } from "@/lib/oidc/routes";
-import { OIDC_MOUNT_PATH, getIssuer, getPublicOrigin } from "@/lib/oidc/tokens";
+import {
+  OIDC_MOUNT_PATH,
+  getIssuer,
+  getPublicOrigin,
+} from "@/lib/oidc/issuer-urls";
 import { getRegisteredRedirectOrigins } from "@/lib/oidc/clients";
 import { isVerifiedCustomDomain } from "@/lib/oidc/custom-domains";
 import { getSecureHeaders } from "@/lib/oidc/security";
