+++ src/app/oidc/device/page.tsx
@@ -12,7 +12,7 @@ import {
   issuerMatchesExpected,
   thirdPartyInitiateSkipCookieName,
 } from "@/lib/oidc/third-party-initiate-login";
-import { getIssuer } from "@/lib/oidc/tokens";
+import { getIssuer } from "@/lib/oidc/issuer-urls";
 
 type SearchParams = Record<string, string | string[] | undefined>;
