+++ src/app/oidc/interaction/page.tsx
@@ -5,7 +5,7 @@ import { IncomingMessage, ServerResponse } from "http";
 import { Socket } from "net";
 import { authOptions } from "@/lib/next-auth-options";
 import { getProvider } from "@/lib/oidc/provider";
-import { getPublicOrigin } from "@/lib/oidc/tokens";
+import { getPublicOrigin } from "@/lib/oidc/issuer-urls";
 import { resolveAppBrandingByClientId, shouldUseWhiteLabelBranding } from "@/lib/oidc/branding";
 import { resolveHostContext } from "@/lib/oidc/host-resolution";
 import { checkAppAccess } from "@/lib/oidc/app-access";
