+++ src/lib/oidc/access-token-verify.ts
@@ -1,59 +1,18 @@
 import * as jose from "jose";
-
-export const OIDC_MOUNT_PATH = "/api/v1/oidc";
-
-function trimTrailingSlash(value: string): string {
-  return value.replace(/\/+$/, "");
-}
-
-function ensureHttpsForProduction(url: string): string {
-  try {
-    const u = new URL(url);
-    const isLocalhost =
-      u.hostname === "localhost" ||
-      u.hostname === "127.0.0.1" ||
-      u.hostname.startsWith("192.168.") ||
-      u.hostname.endsWith(".local");
-    if (!isLocalhost && u.protocol === "http:") {
-      u.protocol = "https:";
-      return u.toString();
-    }
-  } catch {
-    /* ignore */
-  }
-  return url;
-}
-
-export function getPublicOrigin(): string {
-  const raw = process.env.NEXTAUTH_URL || "http://localhost:3001";
-  return trimTrailingSlash(ensureHttpsForProduction(raw));
-}
-
-export function getIssuer(): string {
-  const configured =
-    process.env.OIDC_ISSUER || process.env.NEXTAUTH_URL || "http://localhost:3001";
-  const normalized = trimTrailingSlash(ensureHttpsForProduction(configured));
-  return normalized.endsWith(OIDC_MOUNT_PATH)
-    ? normalized
-    : `${normalized}${OIDC_MOUNT_PATH}`;
-}
-
-export function getCanonicalIssuer(): string {
-  return getIssuer();
-}
+import { getCanonicalIssuer } from "./issuer-urls";
 
 /**
  * Verify a JWT access token issued by the OIDC provider.
  *
  * Validates the signature against the local JWKS, checks issuer, and
  * verifies the audience matches.
- * 
+ *
  * This function always validates against the canonical issuer, even when
  * custom domains or future per-tenant issuers are in use. Token validation
  * is issuer-centric and centralized for security.
  */
 export async function verifyAccessToken(
-  token: string
+  token: string,
 ): Promise<jose.JWTPayload | null> {
   try {
     const issuer = getCanonicalIssuer();
@@ -78,7 +37,7 @@ export async function verifyAccessToken(
  */
 export async function verifyAccessTokenWithIssuer(
   token: string,
-  expectedIssuer: string
+  expectedIssuer: string,
 ): Promise<jose.JWTPayload | null> {
   try {
     const canonicalIssuer = getCanonicalIssuer();
