+++ src/lib/oidc/device-token-exchange.ts
@@ -13,7 +13,8 @@ import { hasScope } from "@/lib/auth";
 import { validateClientSecret } from "@/lib/oidc/clients";
 import { normalizeUserCode } from "@/lib/oidc/device";
 import { approveDeviceCodeForAccount } from "@/lib/oidc/device-approval";
-import { getIssuer, verifyAccessToken } from "@/lib/oidc/tokens";
+import { verifyAccessToken } from "@/lib/oidc/access-token-verify";
+import { getIssuer } from "@/lib/oidc/issuer-urls";
 import { TokenExchangeError } from "@/lib/oidc/token-exchange";
 import { findOrCreateAppEndUser } from "@/lib/billing";
 import { createCorrelationId, writeAuditLog } from "@/lib/audit";
