+++ src/lib/oidc/index.ts
@@ -1,5 +1,6 @@
 export * from "./jwks";
-export * from "./tokens";
+export * from "./issuer-urls";
+export * from "./access-token-verify";
 export * from "./clients";
 export * from "./account";
 export { getProvider } from "./provider";
